[#59] 컴파일 함수 단위 분리
